requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
clip==0.2.0
clip-by-openai==1.1
cmake==3.22.5
cycler==0.11.0
dataclasses==0.8
Django==3.2.14
django-cors-headers==3.10.1
djangorestframework==3.13.1
dlib==19.24.0
ftfy==6.0.3
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
lpips==0.1.4
matplotlib==3.3.4
ninja==1.10.2.3
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.24
requests==2.27.1
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
zipp==3.6.0
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
clip==0.2.0
clip-by-openai==1.1
cmake==3.22.5
cycler==0.11.0
dataclasses==0.8
Django==3.2.14
django-cors-headers==3.10.1
djangorestframework==3.13.1
dlib==19.24.0
ftfy==6.0.3
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
lpips==0.1.4
matplotlib==3.3.4
ninja==1.10.2.3
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.24
requests==2.27.1
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.10
uWSGI==2.0.20
wcwidth==0.2.5
zipp==3.6.0
